+++ atlasdb-cassandra/src/main/java/com/palantir/atlasdb/keyvalue/cassandra/CassandraKeyValueServiceImpl.java
@@ -188,7 +188,6 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
     private final Optional<LeaderConfig> leaderConfig;
     private final HiddenTables hiddenTables;
 
-    private final SchemaMutationLockTables lockTables;
     private final UniqueSchemaMutationLockTable schemaMutationLockTable;
 
     private ConsistencyLevel readConsistency = ConsistencyLevel.LOCAL_QUORUM;
@@ -306,7 +305,7 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
         this.leaderConfig = leaderConfig;
         this.hiddenTables = new HiddenTables();
 
-        this.lockTables = new SchemaMutationLockTables(clientPool, config);
+        SchemaMutationLockTables lockTables = new SchemaMutationLockTables(clientPool, config);
         this.schemaMutationLockTable = new UniqueSchemaMutationLockTable(lockTables, whoIsTheLockCreator());
 
         this.queryRunner = new TracingQueryRunner(log, tracingPrefs);
@@ -2104,28 +2103,6 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
     @Override
     public boolean performanceIsSensitiveToTombstones() {
         return true;
-    }
-
-    /**
-     * Does not require all Cassandra nodes to be up and available, works as long as quorum is achieved.
-     */
-    @Override
-    public void cleanUpSchemaMutationLockTablesState() throws TException {
-        Set<TableReference> tables = lockTables.getAllLockTables();
-        java.util.Optional<TableReference> tableToKeep = tables.stream().findFirst();
-        if (!tableToKeep.isPresent()) {
-            log.info("No lock tables to clean up.");
-            return;
-        }
-        tables.remove(tableToKeep.get());
-        if (tables.size() > 0) {
-            cassandraTableDropper.dropTables(tables);
-            LoggingArgs.SafeAndUnsafeTableReferences safeAndUnsafe = LoggingArgs.tableRefs(tables);
-            log.info("Dropped tables {} and {}", safeAndUnsafe.safeTableRefs(), safeAndUnsafe.unsafeTableRefs());
-        }
-        schemaMutationLock.cleanLockState();
-        log.info("Reset the schema mutation lock in table [{}]",
-                LoggingArgs.tableRef(tableToKeep.get()));
     }
 
     private static class TableCellAndValue {